requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
attrs==21.2.0
    # via
    #   jsonschema
    #   pytest
colorama==0.4.4
    # via copier
copier==5.1.0
    # via -r requirements.in
importlib-resources==5.4.0
    # via jsonschema
iniconfig==1.1.1
    # via pytest
jinja2==2.11.3
    # via copier
jsonschema==4.3.2
    # via -r requirements.in
markupsafe==2.0.1
    # via jinja2
packaging==20.9
    # via
    #   copier
    #   pytest
pathspec==0.8.1
    # via copier
pluggy==1.0.0
    # via pytest
plumbum==1.7.2
    # via
    #   -r requirements.in
    #   copier
py==1.11.0
    # via pytest
pydantic==1.8.2
    # via copier
pyparsing==3.0.6
    # via packaging
pyrsistent==0.18.0
    # via jsonschema
pytest==6.2.5
    # via -r requirements.in
pyyaml==5.4.1
    # via
    #   copier
    #   pyyaml-include
pyyaml-include==1.2.post2
    # via copier
regex==2020.11.13
    # via copier
toml==0.10.2
    # via pytest
typing-extensions==4.0.1
    # via pydantic
zipp==3.6.0
    # via importlib-resources
